Including remaining files that I forgot it

+++ core/product/product.go
@@ -5,5 +5,5 @@ type Product struct {
 	Name        string  `bson:"name"`
 	Description string  `bson:"description"`
 	Price       float64 `bson:"price"`
-	Quantity    float64 `bson:"quantity"`
+	Quantity    int32   `bson:"quantity"`
 }
